Add interpreters for Reader and State based on HList
Normally, you would need to compose tons of runReader and runState
functions for each type in your effect. runReaders and runStates
are new functions which allow you to pass in an Hlist
and automatically interpret all of the values inside. For reader
or state heavy effects, this should cut down on boilerplate.

The Readers and States types are designed to decrease
boilerplate in type signatures in similar circumstances.