Add interpreters for Reader and State based on HList (#18)
* Add interpreters for Reader and State based on HList

Normally, you would need to compose tons of runReader and runState
functions for each type in your effect. runReaders and runStates
are new functions which allow you to pass in an Hlist
and automatically interpret all of the values inside. For reader
or state heavy effects, this should cut down on boilerplate.

The Readers and States types are designed to decrease
boilerplate in type signatures in similar circumstances.

* Add runSeveral as an abstraction over run*s

The run*s functions are now implemented as partial application on
runSeveral. RunConstInputs was also added.

* Export TypeMap and TypeConcat to go with runSeveral

Users may want to create their own specializations for runSeveral.
If they do, they may want to give their specializations type
signatures. Doing that requires having those type families so this
adds them to the export list.

* Remove specialized functions

* Change name to Several