Continuation Effects (#34)
This adds the effects Cont and Shift, for abortive and delimited continuations respectively.

These effects work by being parametrized by s and m, where s is the final result type, and m is the final monad type.
Polymorphic code with these effects are intended to be polymorphic in m, and may choose to be polymorphic in s. This way, you can get the traditional callCC/shift/reset operations. If you specialize s, you also get access to abort, which lets you end a computation early.

Unfortunately, the interpreters for these are wonky. The only way to transform
Union r (Sem (Cont s m ': r)) x to Sem r x is to "replace" any higher-order use of values of type Sem (Cont s m ': r) within the union with a pure value of Sem r. The effect of this is that any continuation you get with callCC/shift is unaffected by higher-order operations such as local and censor, and listen on a continuation will just return mempty.

Still, it's pretty neat that this is at all even possible, considering ContT is not a functor on monads.